+++ CHANGELOG.rst
@@ -1,13 +1,21 @@
 Change log
 ================================================================================
 
+0.7.10 - 16.08.2020
+--------------------------------------------------------------------------------
+
+**Updated**
+
+#. `#393 <https://github.com/moremoban/moban/issues/393>`_: Rendered content
+   output to stdout once
+
 0.7.9 - 06.08.2020
 --------------------------------------------------------------------------------
 
 **Updated**
 
 #. `#390 <https://github.com/moremoban/moban/issues/390>`_: single render action
-   will print to stdout by defafult
+   will print to stdout by default
 
 0.7.8 - 09.06.2020
 --------------------------------------------------------------------------------